+++ src/cli/formats/tta_seek.c
@@ -71,7 +71,7 @@ seektable_add(
 	}
 	assert(st->table != NULL);
 
-	if UNLIKELY ( value > UINT32_MAX ){
+	if UNLIKELY ( value > (size_t) UINT32_MAX ){
 		warning_tta("%s: frame %zu: seektable overflow",
 			outfile_name, framenum
 		);
